comics_project/requirements.txt
Django==3.0.5
django-extensions==2.2.9
Pillow==7.1.1
psycopg2-binary==2.8.4
pytesseract==0.3.3
djangorestframework==3.11.0
django-filter==2.2.0
celery==4.4.2
requests==2.23.0
